Change ingress delay for KSZ9893R Ethernet switch on fsimx8mm/mn
When using the KSZ9893R Ethernet switch on PCoreMX8Mx PCB on
i.MX8M-Nano,
the network transfers are very unstable. Setting a 1.5ns delay on
ingress
traffic solves the problem. This means that the timings on i.MX8M-Mini
are most probably also very tight and we should also use the delay
there.
Tests show that data transfer works equally good with the delay there.

So add the delay to both files, fsimx8mn.c and fsimx8mm.c.

Settings for the PHY are only meaningful when using the PCoreMX8MM PCB,
so remove the code in board_phy_config() in case of BT_PICOCOREMX8MX.