Allow optional board-fdt in BOARD-CFG files
The device tree name for the U-Boot device tree is derived from the
board-name in the BOARD-CFG by making the name lower case. However in
some cases it might be useful to have a different device tree name.

This commit adds the possibility to have an optional board-fdt entry
in the board-cfg section of the BOARD-CFG file. Then this device tree
is used instead.

Example:

/ {
	board-cfg {
		...
		board-name = "PicoCoreMX8MN-DDR3L";
		board-rev = <120>;		/* Rev 1.20 */
		board-fdt = "my-special-fdt";
                ...
	};
};

Then SPL would use the "my-special-fdt" device tree for this device to
start U-Boot.

Remark:
Please note that this entry is only used for the U-Boot device tree.
The Linux device tree name is still determined by executing U-Boot's
environment variable set_bootfdt, where the default is still based on
the board-name.