On fsimx8mn, disable code to detect if running on secondary SPL
The bit in SRC gpr10 that shows if we are running on Primary or Secondary
Boot Image on i.MX8M-Mini does not exist on i.MX8M-Nano. Comment out the
code that does the detection. We still have to figure out how we can
determine this on i.MX8M-Nano.