Fix imx6 display
When updating the uboot to the imx version, some files where kept the
F&S way, like ipu_common , because we want to keep as many of the F&S
changes as possible we decided to. Other files, like ipu_disp where
updated to the imx variant. This lead to a mismatch between certain
definitions, like g_ldb_clk beeing a array in ipu_disp.c and a pointer
in ipo_common.c, which lead to uboot crashes.

We now use the F&S code for ipu_disp.c too.