Add env import/export cmd
NXP undefined some UBoot commands probably for security reasons. For our
non-secure UBoot this should not matter. For the secure UBoot more
research should be done anyway to see which commands and functions are
critical for security.

In UBoot version 2020.04 these commands will be disabled by NXP in the
defconfig.