In mxs-nand-fus.c, set and use ecc.total/ecc.bytes correctly
Our code simply used ecc.bytes as the total number of ECC bytes for the
whole page. However ecc.bytes is now only meant for one ECC step, and there
is a different value ecc.total that holds the total number of ECC bytes.

Change the code to set ecc.bytes and ecc.total correctly.

However there is a problem. The BCH hardware on i.MX CPUs may use ECC
sums per step that are not a multiple of bytes. For example when using
ecc_strength = 62 and two steps with 1024 bytes each (gf=14), the total
number of ECC bits for one step is 62*14 = 868 bits, which is 108.5 bytes.
So it is impossible to give a correct value for ecc.bytes, just for
ecc.total. But ecc.total is computed in nand_base.c::nand_scan_tail() as
ecc->steps * ecc->bytes. So this value would always be wrong. Change the
code in nand_scan_tail() to only compute ecc->total anew if it is still
zero at this time. If it is non-zero, it is assumed that the value was
already computed in the attach() function of the specific driver.