Fix bootcmd-fastboot override when booting from USB
When booting from USB, e.g with MFG tool the environment variable
bootcmd gets overridden with 'fastboot 0'. We need this for i.MX8,
because we have to fastboot into Linux for writing the Nboot.
But for i.MX6 there is no fastboot activated, so the fastboot cmd isn't
known.
Overriding the bootcmd here breaks the boot command. So we include this
override only if fastboot is activated in the configs.