Add Ethernet support for efusA9r2
EfusA9r2 needs a RX/TX delay for the RTL8211F PHY. The delay will be set
in phy configuration, if interface mode is set to
PHY_INTERFACE_RGMII_MODE_RGMII_ID, for all the other F&S boards this is
not necessary. Add interface mode configuration settings for this in
fsimx6.c.

Unfortunately the Freescale driver set the interface mode to
PHY_INTERFACE_MODE_RGMII. So extend the driver to allow
configuration on the init call. We modify fecmxc_initialize_multi()
for this and add a version fecmxc_initialize_multi_if_mode() that adds
the interface mode. This requires including phy.h to netdev.h so
that it is available in the board support files.