Improve fsimx6 configuration u-boot size
Due to the reason that U-Boot exceeds the maxium size of 512 KB, we have
to shrink our configuration for fsimx6.
On the one hand the new board 'efusA9r2' has the number 29. Therefore we
have 21 empty entrys in our struct board_info. To save this space we
subtract the board_type number, if it´s equal or bigger than 29 in our
common function fs_board_get_type. Because we have reservered some space
in N-Boot for 29, 30 and 31. This numbers can be i.MX6 boards in future.
On the other hand we remove the U-Boot command line tool CONFIG_CMD_I2C.
This is only useful for debugging purpose and it´s not necessary for the
U-Boot functionality and internal U-Boot I2C commands.