In mxs_nand_spl.c, actually fail in case of non-correctable ECC errors
The return code of nand_chip.ecc.read_page() only returns an error if the
page could not be read due to I/O errors. But it does not return an error
code if the ECC failed. This can only be seen in the ecc_stats.

So in mxs_read_page_ecc(), actually look at the ecc_stats to see if error
correction failed and do fail in these cases, too. Return the proper error
code instead of -1, and also return this error code in nand_spl_load_image().