Prepare NAND on fsimx8mm for device tree support
Prepare fsimx8mm to be able to use NAND via device tree. It requires setting
CONFIG_MTD and CONFIG_NAND_MXS_DT to do so.

- Add nodes dma_apbh and gpmi to fsl-imx8mm.dtsi.
- Add alias, node gpmi and pad settings for gpmi to picocoremx8mm-ddr3l.dts
  and picocoremx8mm-lpddr4.dts.
- Only use board_nand_init() in fsimx8mm.c/fsimx8mn.c if CONFIG_NAND_MXS_DT
  is not set.

On fsimx8mn, a lot of device tree code is still missing in this U-Boot
version, so it does not make sense to change these device trees yet.
Because of this we also keep the non-device-tree version on fsimx8mm
active until the regular support will be added in a future version of
U-Boot. This means fsimx8mm and fsimx8mn will work in the same way for now.