Bring back NXP fastboot version
When switching to u-boot-2018.07, we reverted the fastboot code to
U-Boot. This commit brings back the NXP extension for fastboot
with acmd and ucmd among others. The code is in drivers/fastboot/fb_fsl
with some modifications in drivers/fastboot. This version is backported
from NXP's u-boot-2019.04.

Change fsimx8mm_defconfig and fsimx8mn_defconfig to use fastboot.