Add "is_usb_boot" function for platform fsimx6sx/ul
We dont want to use the common "is_usb_boot" function, which returns
true if we are e.g. transfer an U-Boot via NetDCU-USB-Loader in N-Boot
and execute the image. We booted from fuse so fuse should be only thing,
which should be checked, and not some flags. Therefore we return always
false.