Improve mx6 mmc_get_env_dev/mmc_get_env_part
By default the code just runs mmc_get_boot_dev, if this fails (-1) the
second function which can be indiviual programmed is not called anymore.
This is not correct in our opinion. The board should have the
possibility to decide by itself if it can recongnize what kind of type
it is, if this fails then the next mechanism should try to detect it.
Therefore we changed the way how to detect it.