Have common trained_csr data for i.MX8M DDR RAM timings
The array ddr_ddrphy_trained_csr[] in all RAM timings is always the same.
This is why NXP has added file drivers/ddr/imx/imx8m/ddrphy_csr.c that
holds a global instance of the data. As a consequence, the data can be
removed in all RAM timings. This reduces the size of each RAM timing by
5744 bytes.

Please remember to remove this part in all new RAM timings that are
created in the future.