Move dram_init() to arch/arm/mach-imx/imx8m/soc.c
The code for rel_imx_5.4.3_2.0.0 added own versions of dram_init() and
dram_init_banksize() to arch/arm/mach-imx/imx8m/soc.c. This change has
impact on other F&S code, so we moved this change to this separate commit
so that it can be found in the history.

In the past we had our own dram_init() function in fs_board_common.c and
we used the default implementation of dram_init_banksize() that simply
created one bank entry for the whole RAM area. The new code for i.MX8M
may create more bank entries, especially if an opTEE is involved. Then
the opTEE RAM region has to be excluded to avoid speculative RAM accesses
to opTEE that would allow side channel attacks.

If opTEE is found, the first RAM region starting at 0x4000_0000 may be
split up in two parts, one part from the start of RAM to the beginning
of opTEE and a second part from the end of opTEE to the end of RAM.
If opTEE happens to be located at the end of the RAM, this second part
can be omitted.

Independent from this, there may be a second physical RAM bank starting at
address 0x1_0000_0000. This will be added as second or third entry
respectively.

The original NXP code creates a memory table with fixed entries and only
updates the table in enable_caches() if an opTEE was detected. However we
never have a fix RAM size, we read it at runtime from the BOARD-CFG. So we
update the memory table in any case before actually enabling the caches.

Instead of dram_init() that we used in fs_board_common.c, we can use the
new function board_phys_sdram_size() now that simply needs to return the
full RAM size. We do not have to care about opTEE anymore, this is
automatically handled in the new dram_init() in soc.c.