Move SPL clock initialization from clock_imx8mm.c to spl.c
On i.MX8MM and i.MX8MN, some clocks that are typically used in SPL were
initialized in clock_init() in arch/arm/mach-imx/imx8m/clock_imx8mm.c
before. But because the generic code does not know what devices are
actually needed on a specific board, the clocks to all UART devices,
all USDHC devices, all ECSPI devices and to NAND were started. This
were most probably enable more clocks that are actually needed.

Add separate init functions for these clocks and let the board specific
start-up code in board/<manufacturer>/<arch>/spl.c decide which clocks
to enable. This is a change that comes in from NXP rel_imx_5.4.3_2.0.0,
but we moved it to a separate commit so that it is better visible in the
history.

The following functions were added:

- init_uart_clk(index)
- init_clk_usdhc(index)
- init_clk_ecspi(index)
- init_nand_clk()

The index is zero-based, e.g. UART1 has index 0 and USDHC3 has index 2.

In spl.c for fsimx8mm and fsimx8mn, start UART clock in config_uart(),
and start USDHC and/or NAND clock in fs_spl_init_boot_dev().