Fix disassembler output
The disassembler listings shall only be built if CONFIG_DISASM,
CONFIG_SPL_DISASM and/or CONFIG_TPL_DISASM are set. And of course only
if the appropriate target is built. This makes it difficult to set the
.dis files as part of the global target list ALL-y, because it is
impossible to decide at the top-level of the Makefile whether the target
is built or not. For example u-boot-spl is built as prerequisite for
many other targets and it is impossible to define a condition when
to add it to ALL-y and when not.

So instead of adding them to ALL-Y, add the disassembler command to the
rules that actually build u-boot, u-boot-spl and u-boot-tpl.bin. Then
the disassembler listing is created whenever the original target is built.

Also bring back CONFIG_SPL_AUTOBUILD and CONFIG_TPL_AUTOBUILD that
unintentionally got lost in one of the previous merges. Only if these
CONFIGs are set, the SPL/TPL is automatically built when building
U-Boot proper. This is the default behavior in U-Boot. However on F&S
boards based on any i.MX8 CPU, SPL is only built as part of NBoot,
not as part of regular U-Boot. So CONFIG_SPL_AUTOBUILD is unset in all
fsimx8* defconfigs.