Improve imx7ulp bootaux command
The imx7ulp bootaux command showed "Aux core still running, abort!" if
the core was already started. This seems like an error message but
actually the M4 core is supposed to keep running when the A7 core gets
rebooted, as it is the main core.
Correct the output message so it is more understandable.