Improve imx7ulp reset cause handling
Use the global gd->arch.reset_cause variable to save the reset cause
once and read it out in the get_reset_cause() function. This way we can
always clear the reg_ssrs register right away and still have the reset
cause saved.
This is also done for the fsimx6* boards.

Also add WARM-CORE0 as reset cause if the Cortex-M4 core rests the
Cortex-A7.