Fix fsimx7ulp reloction error
The global struct fs_nboot_args nbootargs was used before relocation so
there where not variables available. This corrupted the relocation table
wich caused some uboot code to get increased by the relocation offset
although it should not be. In this specific case fuse prog caused a cpu
reset but it could have had any effect.

We now initialise pargs directly so there is no need for the global
struct.