Reduce diff to rel_imx_5.4.70_2.3.0
Because of the many merges and inbetween versions we have added, there
have accumulated some minor differences between our F&S version and the
original NXP version. With this commit, we try to get as close as possible
to the original version again. This will minimize conflicts in the future.

Typical differences are:

- Different indendation or spacing.
- Differences in files not relevant for the F&S boards (e.g. EVK
  boards from NXP).
- Changes that were skipped in the past.

All these differences can be removed.