Call fix_fdt() earlier in board_f.c
The board_f code reserved a memory region for the fdt in reserve_fdt()
and then called fix_fdt(). However this function typically modifies the
device tree which may result in a bigger device tree. But a bigger device
tree exceeds the reserved region, causing variables behind the region to
be destroyed when the fdt is actually copied there later in reloc_fdt().

The solution is to call fix_fdt() earlier, before reserve_fdt(), so that
the memory region is reserved based on the new size already. Then it will
always fit in reloc_fdt(). Later u-boot versions will do this, too.