Remove CONFIG_SYS_STACK_SIZE from F&S configs
CONFIG_SYS_STACK_SIZE was only used in the F&S version of memtest.
The variable should set the maximum size of the stack. However the
stack is never actually checked and can grow indefinitely. So the
value was always only a guessed size. We can also guess a value in
cmd/mem.c directly, without the need for an CONFIG setting. So use
16KB there and remove all these macros from the .h files of F&S
board configs.