Fix env device for fsimx8mp
For some reason, env_get_ulong("mmcdev", 10, CONFIG_SYS_MMC_ENV_DEV) in
arch/arm/mach-imx/mmc_env.c returns 0 on fsimx8mp, despite the fact that
CONFIG_SYS_MMC_ENV_DEV is 2 (MMC3). However the whole construct is strange.
We are just about to determine where the environment is located. So it does
not make sense to read an enviroment variable. We cannot read something from
that we do not know the location yet. So remove the env_get_ulong() and
use CONFIG_SYS_MMC_ENV directly. This is how mainline U-Boot does this,
too. Only NXPs version uses the env_get_ulong().

Because the PicoCoreMX8MP boots from MMC3, we also have to add this to
the list of known locations in env_get_location() in fsimx8mp.c.