Disable ethernet before booting linux for fsimx8mm
Ethernet was not disabled before booting linux for fsimx8mm so far, like
with the fsimx6 architectures.
So if ethernet was disabled by default in linux the phy still consumed a
lot of power.
Add the mdio_shutdown_all() to the fsimx8mm preeboot_os hook so it gets
disabled as expected.