Improve fsimx8mm dram timings
Because DRAMs with similar parameters produce the same dram timings, we
rename the dram timings to more generic names, which describe the dram
type, size, speed,number of channels and number of chip selects.
To save space, we strip the dram tpye from the file names in the
Makefile, because it is already given in the dram-type property.

We have two RAM timings for 1GB lpddr4 with different speed settings.
The old RAM timings are slower but consume less power and produce less
heat.
We update all boards, except the old customer specific boards, to the
new faster timings.

The dram timing update of the customer specific boards have to be
discussed with the customer or they might get problems with power
consumption or heating.