Remove XTAL-OFF setting for ALDPS in realtek driver
For boards without a quartz on the realtek phy, the XTAL-OFF setting
should not be enabled because the reference clock will be shut down. In
Linux this should be okay as the clocks will be reinitiated for a link
up, but in UBoot this seems to break the ethernet connection completely.
This could be because of the interrupt handling in Linux, which is not
present in UBoot.