Revert F&S changes in iomux.c and console.c
Long ago, the stdio, console and serial stuff in U-Boot was very bad and
did not work well when using several serial ports and console muxing. So
F&S changed the stdio handling quite considerably to make it easier to
switch between different serial ports. Since then, mainline U-Boot has
partly cleaned up the code in a similar way, so F&S has returned to the
original code in many places. However as we have not used console muxing
for quite a while, the appropriate code in cmd/console.c and common/iomux.c
was not cleanly reverted.

On our i.MX8M* boards, CONFIG_FASTBOOT_UUU_SUPPORT is currently required to
install the system software, especially when SPL has to be written to NAND.
This pulls in console muxing and triggers the non-working code in the files
above. Fix the code by also going back (more or less) to the original U-Boot
code -- even though the code in iomux_doenv() is extremely ugly.