Enable CONFIG_FASTBOOT_UUU_SUPPORT on fsimx8mm/mn/mp
The i.MX8M* boards need UUU support for now to install the system software,
especially if SPL has to be written to NAND by kobs. Most probably, we will
change this in the future as this is a big security hole, but for now we
still need it. However setting CONFIG_FASTBOOT_UUU_SUPPORT selects quite a
lot of stuff like console muxing and GPT stuff that we do not want.

Change the select entries to imply entries in drivers/fastboot/Kconfig.
Then we can still leave out the unnecessary stuff.

Enable CONFIG_FASTBOOT_UUU_SUPPORT accordingly in fsimx8mm_defconfig,
fsimx8mn_defconfig and fsimx8mp_defconfig.