Add no-uboot-override to fs_fdt_set_val()
At some places, overriding a device tree value for Linux by U-Boot can be
prohibited by adding "no-uboot-override" to the node. However this was
handled separately in every case, resulting in rather bloated code.

Make this more general by adding this query to the basic function
fs_fdt_set_val() in fs_fdt_common.c. Then we can set this property in
*all* nodes that may get values overridden.

As a result, all the places where this was handled separately, can now
be written much simpler. This was the linux,cma setting and the settings
for alert and critical temperature.

Also fix a compiler warning in case of fsimx8mp because the temp_range
value is not used yet.