Update fsimx8x_defconfig and fsimx8.h
Change fsimx8x_defconfig to be more similar to fsimx8mm/mn/mp_defconfig
and provide values for CONFIG macros that were introduced in the meantime.

Also update fsxim8x.h, especially the head comment with the eMMC layout.
Drop the NAND layout, this is not planned at the moment. Some values have
been moved to defconfig in the meantime, so remove them here,