Fix alignment violation for zImage in common/image.c
When checking for IH_ZMAGIC in function genimg_get_format(), img_addr was
casted to a ulong pointer. However on 64 bit CPUs, ulong is 64 bit, which
causes an alignment violation if img_addr ends in 0x0 or 0x8 as final digit.
IH_ZMAGIC is a 32 bit value, so using ulong is wrong anyway. Change this to
a u32 pointer instead.

Instead of excluding IMX8M, IMX8MM and other specific architectures,
exclude ARM64 in general, as ARM64 does not support zImage format at all.