Fix SDP timeout errors on fsimx8x
When NBoot download is finished (OUT report 2), the board will answer
with an IN report 3 (HAB status). Then it sends an IN report 4 (transmit
status) right after. This transmission was not acknowledged correctly by
the UUU tool on the host side. Research showed, that UUU always expects
65 bytes when loading these reports. But report 3 only sends 5 bytes.
Unfortunately, older Windows USB drivers (e.g. on Win 7) will not detect
this correctly and take bytes from the next report to fill up the first
report. This causes the second report to be incomplete and UUU does not
see the second report, causing a timeout error.

By setting both reports, 3 and 4 to the length of 65 bytes (1 byte report
ID, 64 bytes data), the consolidation of the messages on the Windows side
can be avoided and UUU sees all reports correctly.