Improve functions from arch/arm/mach-imx/checkboot.c
The functions from checkboot.c were originaly written for use with imx6
Boards. All of those Boards are 32 bit Systems. Thus pointers were
written with uint32_t variables. On imx8m Boards, which are 64 bit
Systems, this will lead to compiler warnings when they are cast to
void*.

To prevent this, change the uint32_t to uintptr_t. This types length is
based on the Systems address width, and therefore works on both imx6 and
imx8m Boards.