Add support for op-tee
Op-tee is a trusted execution environment with support on Linux. Add
changes so op-tee can be loaded. op-tee support will only activate when
using a special defconfig. These condigs activate options that prepare
the environment for op-tee and split the DRAM into two virtual Banks
since op-tee needs to be invidible for the User environment.

For i.MX8MM it was also necessary to change the predefined op-tee load
address, since no current Board has enough RAM to support the former
address.

Adjust the op-tee address so it runs at the same address as on other
i.MX8M Architectures, this has the benefit that it is able to run on any
board with two virtual DRAM banks.