Fix OC-RAM Layout for fsimx8mp
Until now we used OC-RAM from 0x910000 to 0x98FFFF. The space from
0x90000 to 0x90FFFF was reserved for the boot-ROM. The reference manual
of the i.MX8M Plus states on the other hand that boot-ROM needs to go
from 0x90000 to 0x917fff. This can cause problems if ROM or HAB APIs are
used since the processor uses this space to save data.

The ATF Area is half as big on fsimx8mp as on other fsimx8m variants.
This will cause the authetication for the DRAM-FW to fail.

Change the OC-RAM to take secure boot into account.