Switch off clocks in imx8_set_parent_mux() before switching parents
Switching clock parents via SCU is only possible if the clock is not
enabled. This implies that any call to imx8_set_parent_mux() assumes
that clocks are disabled. When only looking at U-Boot itself, this is
true, because the devices have not been activated yet when this function
is called. However we also have to consider SPL. Some devices (like
eMMC) may still be active from SPL and therefore have an active clock
here. So to be sure, disable the clock first before switching the clock
parent. The clock will be activated again later in the code.