Keep the USB product ID 0xb4a4 during SDP download
New U-Boot would set USB idProduct to 0x0151, which seems to be valid
for newer versions of uuu. However we still use SDPV download mode, even
on boards that are designed to use SDPS, so we need 0xb4a4. So keep the
old code in arch/arm/mach-imx/spl.c in function g_dnl_bind_fixup().