Add mcu_rdc settings to fsimx8mp and fsimx8mn device trees
On i.MX8MP and i.MX8MN architectures, the Resource Domain Controller (RDC)
settings are copied to OCRAM_S in board_spl_fit_post_load(). There is a
Start Configuration that is set when the MCU (Cortex-M) is started, and a
Stop Configuration that is set when the MCU stops.

The default addresses in OCRAM_S are (on both i.MX8MP and i.MX8MN):

  0x186000 Start Configuration
  0x187000 Stop Configuration

Handling of these configurations is done in an up-to-date ATF version that
passes these settings to the RDC as part of the IMX_SIP_SRC_MCU_START call.

So update all fsimx8mp and fsimx8mn device trees to contain such Start
and Stop Configurations.

ATTENTION:
These configurations are taken from the imx8mp-evk and imx8mn-evk device
trees. They are most probably not fully suitable for F&S boards and need
to be adjusted when Cortex-M will be used in the future.