Improve Uboot versioning
We want to add our own git tags into the uboot version string. The
default implementation did not print tags because the version is already
given in the make file.
This is not true for F&S versions so we add out git tags to the build.

The order should be like this:
1. If there is a .scmversion in the root-dir use its content.
2. If there is a Git tag use it. (The tag has to be annotated)
3. If there is a Git hash use it.