Add display flag vcfl_inv to fs_disp_common for fsimx6/sx
Customer boards may also need an inverted output to enable the VCFL on
a display. Similar to bl_inv, add a flag vcfl_inv to be able to define
the polarity of the VCLF signal in the environment variable dispflags.

Example:

  setenv dispflags bl_inv=0,vcfl_inv=1

Also update the display flags mask setting in fsimx6.c, fsimx6sx.c,
and fsimx6ul.c.