Replace HAVE_BOARD_CFG with CONFIG_FS_BOARD_CFG, add CONFIG_FS_NBOOT_ARGS
In board/F+S/common, there needs to be some different code depending on
whether we are on a legacy board with old (closed source) NBoot (Vybrid,
i.MX6 platforms) or on a newer board with SPL based (open source) NBoot
(i.MX8X, i.MX8M platforms). In the past this was decided by setting a
macro HAVE_BOARD_CFG.

Now we need this information in other places, too. For example there
is an entry added to struct global_data if we are on a board with
BOARD-CFG support. And we will show an additional entry in the bdinfo
command in this case. So switch to a regular Kconfig entry. We also
add an entry for the old variants. This will make it easier in the
future to split fs_board_common.c in three parts: a real common part,
a part with functions only for legacy, nboot_args-baserd boards and
a part with functions only for new BOARD-CFG-based boards.

- Add CONFIG_FS_BOARD_CFG and CONFIG_FS_NBOOT_ARGS to
  board/F+S/common/Kconfig.
- Activate the appropriate setting for i.MX8, i.MX6 and Vybrid boards:
  arch/arm/mach-imx/imx8/Kconfig, arch/arm/mach-imx/imx8m/Kconfig,
  arch/arm/mach-imx/mx6/Kconfig, arch/arm/Kconfig.
- Replace HAVE_BOARD_CFG in all F&S common files (board/F+S/common)
  where it was used.