NBoot: Handle relocated BOARD-CFG in OCRAM
The location, where the BOARD-CFG is stored in OCRAM, is the only address
where SPL and U-Boot proper have to agree on. All remaining addresses are
taken from this BOARD-CFG then (from node nboot-info).

Unfortunately, we had to move this location to a new address on fsimx8mp,
because the old address was still within the region that is used by the
ROM loader, which is important in case of Secure Boot. As a consequence, an
older U-Boot, that still had the old address configured, could not find the
BOARD-CFG anymore, which provides quite some difficulties when updating.

With this commit, U-Boot will search the whole OCRAM for the BOARD_CFG if it
did not find it at the expeceted location. This is possible because the
BOARD-CFG is prepended by an F&S Image Header that can be identified easily.

- Add function fs_image_get_regular_cfg_addr() to fs_image_common.c that
  returns the intended address of the BOARD-CFG in OCRAM.
- Add CONFIG_SYS_OCRAM_BASE and CONFIG_SYS_OCRAM_SIZE to fsimx8mm/mn/mp/x.h
  (and config_whitelist.txt) to hold the location and size of the OCRAM.
  While we are at it, also align some more values to the correct column in
  these header files.
- Add function fs_image_find_cfg_in_ocram() to fs_image_common.c that
  searches the whole OCRAM for the BOARD-CFG if it is not found at the
  regular location. To avoid having to search several times, add entry
  board_cfg to struct global_data to store the final address. From then
  on, fs_image_get_cfg_addr() will return the proper address.
- Add function fs_image_cfg_is_valid() to verify the integrity of the
  BOARD-CFG in OCRAM. For now, only the F&S image header will be checked
  again, but in the future, this should verify at least a checksum, and in
  case of Secure Boot, this should re-authenticate the whole BOARD-CFG.
- Rename fs_spl_setup_cfg_info() in fsimx8mm/mn/mp/x.c to fs_setup_cfg_info().
  Search and verify the BOARD-CFG there. Stop if BOARD-CFG could not be found
  or is corrupted.
- In fs_board_late_init_common(), show the real location of the BOARD-CFG as a
  boot message, add a warning for incompatible NBoot and U-Boot if it is not
  at the expected place.
- Also check the validity of the BOARD-CFG in OCRAM once in the basic fsimage
  command and show a warning if there is some inconsistency. This makes
  function fs_image_get_cfg_addr_check() obsolete, remove it.