Add fs_image_get_cfg_fdt() to get pointer to fdt part of BOARD-CFG
In the past, fs_image_get_cfg_addr() had a boolean argument that told the
function to return a pointer to the BOARD-CFG in OCRAM either with or without
the F&S Header. In fact the second variant without header was meant to get a
pointer to the actual fdt of the BOARD-CFG. However on a board with signed
images, there is also the IVT between F&S header and the fdt, so computing
the fdt pointer by skipping the F&S header is not sufficient then. But then
it is not simply "with or withiut F&S header" anymore.

So let's have two functions now. The original function returns a pointer to
the full image including F&S header now and has no parameter anymore. And
a new function fs_image_get_cfg_fdt() returns a pointer to the fdt, no
matter if signed or not.

In the fsimage command, we also need the fdt offset to a BOARD-CFG that is
somewhere else in RAM. So also add a function fs_image_find_cfg_fdt()
that gets a pointer to the full BOARD-CFG as argument and returns the
pointer to the fdt. This makes function find_fdt_signed() in fsimage.c
obsolete.

Change all references to these functions accordingly.