On fsimx8mm/mn/mp/x, move cfg_info from OCRAM to global_data
To avoid needing to parse the BOARD-CFG several times in U-Boot proper,
all necessary information is copied to a binary data structure cfg_info
rather early. This was located in OCRAM. However there it is vulnerable
to modifications in the same way as the BOARD-CFG itself. Using a regular
variable is also not possible, because it is used too early when regular
variables are not available yet.

So move the structure to the global_data. This can be accessed at any
time. The struct size is not very big, so this should be no problem.

- Change fs_board_get_cfg_info() in fs_bard_common.c accordingly to
  return the pointer to global_data now.
- Move struct cfg_info from fs_board_common.h to a separate global
  include file include/fs_cfg_info.h to avoid haveing to include a
  vendor-specific file in include/asm-generic/global_data.h.