Add more CRC32 options to addfsheader.sh and fsimage.sh
Move padding option from -p/--pad to -a/--align. Move -s/--set option
-p/--param. The p/param is automatically the beginning of the parameter
name then and simplifies parameter parsing considerably.

When using option -c/--crc32, addfsheader now stores the CRC32 of the
image in type[12..15], not in p32[7] anymore. New -s/--secure option
takes the CRC32 of the F&S header, and if -s and -c are both given, the
CRC32 is taken from the header plus the image. This restricts the type
to 11 characters +'\0' then.

Warn user if description exceeds 32 or type exceeds 16 characters (or 11
in case of CRC32).

Change fsimage.sh to be capable of working with the new CRC32 formats.
Also add a column for the well-known flags to the image list: D for
Description, C for CRC32, S for Secure. Handle input arguments more
reliably, especially when in a pipe or input is redirected.