Fix CONFIG_SPL_TEXT_BASE for fsimx8mp_defconfig
The ROM Loader needs more space in OCRAM on i.MX8MP. So there the free
space is at 0x928000 there, not at 0x920000 as on the other i.MX8M CPUs.
This was corrected on the secure boot defconfigs already, this commit
also changes this for the regular fsimx8mp_defconfig.