Activate CRC32 on F&S images for NBoot on fsimx8mm/mn/mp/x
Use the new CRC32 feature of addfsheader on NBoot sub-images. Due to the
fact that the CRC32 is stored at the end of the type entry, the type is
restricted to 11 characters in this case. As all sub-images get a CRC32
now, we have to change BOARD-CONFIGS to BOARD-INFO and DRAM-SETTINGS to
DRAM-INFO to keep this limit.

- A full CRC32 (including F&S header) is used on: SPL, BOARD-CFG, RAM-FW,
  RAM-TIMING, ATF, TEE and NBoot as a whole. These are the image types
  that will be signed in case of Secure Boot. In this case we also store
  the load address, for which the image should be signed for, in p32[7].
- A short CRC32 (over the F&S header only) is used on: BOARD-INFO,
  FIRMWARE, DRAM-INFO and DRAM-TYPE. These are the headers that only group
  a set of sub-images that already have a CRC32/Signature, so that an extra
  full CRC32/Signature makes no sense. But of course we still want to make
  sure that the header itself is intact.
- The EXTRA section with the bash-scripts remain without CRC32 for now. They
  are not directly used in the boot process and therefore can not help in
  improving the boot stability/security.

When saving an image with fsimage save, we currently accept the old
BOARD-CONFIGS and the new BOARD-INFO type. However when loading images in SPL,
only the new variants BOARD-INFO and DRAM-INFO are accepted. As the whole
NBoot is either a new version or an old version, SPL will automatically match
the right set of image types.

Note:
No validation of the CRC32 is currently done in SPL or U-Boot. This will be
added by a later commit.