Add CRC32 checking of new F&S NBoot images to cmd/fsimage.c
New F&S NBoot images use CRC32 to allow consistency checking, too. This
can be handled in a similar way as signed images. So combine the code for
authentication of signed images and CRC32 checksum validation in a common
new function fsimage_validate(). The authentication of signed images is
then handled in the new function fsimage_validate_signed(). Images without
CRC32 (e.g. old NBoot images) are still accepted and are always valid.

The code asssumes that the layout of signed images is unified in the
future, so that the IVT is always directly behind the F&S header. In the
past, there were different layouts for different image types.

Also correctly return CMD_RET_FAILURE and CMD_RET_SUCCESS instead of 1 and 0
in functions called as commands.