Add optional "stored" argument to fsimage fuse/board-cfg
Add optional argument "stored" to fsimage fuse to allow fusing the stored
configuration. In the past we had to load the NBoot image to RAM to be able
to burn fuses for it.

fsimage board-cfg will now check if there is an NBoot image loaded.
Then it will show the contents of this image. Otherwise it will fall
back to the stored configuration. To force using the stored configuration
like in the past, use "stored" as argument instead of an address.

Add function fs_image_get_loadaddr() to help parsing command line arguments
of fsimage subcommands.