Replace command fsimage firmware with fsimage load
The command fsimage firmware was meant to load the stored firmware
to RAM so that it can be inspected with fsimage list. Other than
that it was rather useless. However if we would load SPL and the
BOARD-CFG, too, we could build a minimal NBoot image in RAM, that
even could be saved again with fsimage save. So replace the old
fsimage firmware command with the new fsimage load command which
is now the logical counterpart to fsimage save and makes more sense.
For example if one copy of NBoot is damaged, we could load the
working copy to RAM and save it again, which would result in two
working copies again. This command will also check if both copies
of NBoot still match and have working checksums.

Please note that in case of secure boot, this minimal NBoot image
is not fully functional because the signature of the NBoot image itself
was not stored and thus can not be recreated here. So the loaded image
cannot be saved again on a closed board because of the missing signature.
But this is not intended anyway, so no big problem.

- Rename fs_image_get_info_offs() to fs_image_get_nboot_info_offs(),
  this is what the function actually does.
- Move fs_image_get_storage_size() and fs_image_get_board_cfg_size()
  from fs_image_common.c to fsimage.c. Also move fs_image_get_*_info()
  there and rename to fs_image_get_*_si(). These functions print too much
  info anyway, which is not wanted in SPL. So just use the regular
  fdt_getprop_*() functions in the SPL part in fs_image_common.c.
- Move struct storage_info from fs_image_common.h to fsimage.c, it is only
  needed there now. Prepare for exactly two copies, even if there are more
  values in the nboot-info. Then we can use CPU little-endian start values
  instead of needing to carry the big-endian fdt32_t values all around.
- Add struct flash_info to fsimage.c that holds the parameters of the
  NAND or MMC device to load from.
- Remove do_fs_image_firmware() and fs_image_load_firmware*() functions
  and add do_fsimage_load() and several helper functions instead.

Remark:
fsimage save may not work correctly in this version. It will be reworked
anyway in one of the next commits.