Add SPL dram test for fsimx8mm
The DRAM Test was ported from the linux memory test based on memtester
by Charles Cazabon.
https://github.com/jnavila/memtester

To activate it, enable FS_SPL_MEMTEST_COMMON in the menuconfig.

The board has to be reset after the test, because it needs to activate
the caches in the SPL and we do not know how this affects the Uboot.

i.MX8MM boots the SPL from TCM so we have to apply different cache
attributes.