Fix cmd_nandbcb.c and mxs_nand.c if using 62 or 40 bit ECC
A command hast to return CMD_RET_FAILURE or CMD_RET_SUCCESS, not an error
code. Fix cmd_nandbcb.c accordingly.

When reading or writing the FCB (Firmware Configuration Block), NXP uses
either a 40 bit or a 62 bit ECC (depending on the BCH capabilities). To
have maximum ECC, the data part is divided into eight chunks and 32 bytes
metadata are used in chunk 0. Unfortunately this information was not
available in function mxs_nand_ecc_read_page() in mxs_nand.c, so that
completely wrong status bytes were checked after reading. So no ECC errors
or empty pages could be detected. Add entries metadatasize and chunk_count
to struct mxs_nand_info and set the values accordingly at different places.

When reading only part of a page, nand_base.c uses a page cache. This cache
needs to be invalidated if the ECC is changed. So set nand->pagebuf to -1
in these cases.